+++ crates/rslint_cli/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.3.1] - 2021-10-06
+
 ### Added
 
 - Default support for `.rslintignore` file